scripts: list images and packagegroups from fsl layers
Generate release-notes/source/images.inc and
release-notes/source/packagegroups.inc based on recipes in
meta-fsl-arm, meta-fsl-arm-extra and meta-fsl-demos.

The "heuristic" to determine what is an image/packagegroup is pretty
lame: it just checks if the recipe name contains `image' or
`packagegroup'.

The generated files contain an rst list mapping the image and
packagegroup recipe names to their corresponding descriptions.

The generated files (.inc) have intentionally not been added to this
commit.

Signed-off-by: Mario Domenech Goulart <mario@ossystems.com.br>