script: create doc_utils.py module for commonly used methods
For now it only has "tabularize", a method extracted from
scripts/bitbake-metadata2doc.py that creates reST tables from
bi-dimensional lists.

Also, the method is slightly different from the original, as it applies
rstrip() to the generated table header and rows to avoid polluting the
commit diffs with blank spaces.

Signed-off-by: Lucas Dutra Nunes <ldnunes@ossystems.com.br>
Signed-off-by: Daiane Angolini <daiane.angolini@freescale.com>