scripts: metadata2doc.py: write_soc_pkg: handle imx-test special version
imx-test has a special version (00.00.00) that should be specially
handled when it is the most frequent version alongside another version
(the other version would be more relevant in that case).

It works around errors like:

ERROR: The most frequent versions (['3.14.28-1.0.0', '00.00.00']) for
  imx-test are equally distributed among boards of SoC mx6sl.
  Cannot determine which one to use.

In this case, we pick the version which is not 00.00.00.

Signed-off-by: Mario Domenech Goulart <mario@ossystems.com.br>