Update User Guide's machine-list.inc To Match Release Notes'
The Release Notes have been kept more up todate than the User Guide.
This has resulted in the correct data for machine-list.inc being present
in the release-notes source code but the user-guide version becoming out
of date.

This changeset updates the user-guide/source/machine-list.inc to match
to one currently found in release-notes/source/machine-list.inc.

A future update should be made to automatically base the user-guide
version on the release-notes version so that there are not two identical
sets of information to be maintained.  For now the pragmatic thing to do
is bring the user-guide up to date.