Update Layout Of Support Machines Table In User Guide
The longer Names found in the latest machine-list.inc data will not fit
on the page by default.  This change updates the layout style of the
Supported machines table in the User Guide to match the style used
within the Release Notes.  Using this all the date appears to be
displayed correctly without clipping for a `make latexpdf`