Use release-notes machine-list.inc In user-guide
The user-guide gets little attention compared to the release notes.
Frequently the correct list of supported machines is visible in the
release notes but the user-guide is left very out of date.  This change
switches to using the machine-list.inc file in the release directory in
both documents to help ensure the user-guide stays up to date.

This issue is described in GitHub Issue #28

Signed-off-by: Thomas A. F. Thorne <TafThorne@GoogleMail.com>