Update the SoC list as well as its specific package list
We are updating the SoC list and its specific package list, covering the
new SoC families as i.MX8 series and like.

A corner case has been found which we were not documenting properly the
BSP components due to the versions used on the 3rdparty boards so we now
have a whitelist so we ensure to process only valid ones.

Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>